requirements.txt
fastapi
uvicorn
pillow
pytesseract
googletrans==4.0.0-rc1
python-multipart
